#71 feat: 토큰을 포함하고 있는 해더를 가진 baseQuery 생성
